requirements.txt
# ── Core HTTP & parsing ──────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
cloudscraper>=1.2.71          # Cloudflare bypass for Indeed

# ── Browser automation (LinkedIn / Handshake) ────────────────────
playwright>=1.43.0             # run: playwright install chromium

# ── Scheduling ───────────────────────────────────────────────────
APScheduler>=3.10.4

# ── Config ───────────────────────────────────────────────────────
PyYAML>=6.0.1

# ── Optional (Telegram notifications) ───────────────────────────
requests>=2.31.0               # already listed, used for Telegram API too

# ── Dev / testing ────────────────────────────────────────────────
# pytest>=8.0.0
# pytest-mock>=3.12.0
